Adds TimeControl data class and Serializer
TimeControl stores information relevant to the allotted time for each
player, the increment after each turn, and the type thereof. To conform
to the Identifiable interface, the appropriate properties are added.

To serialize a TimeControl, just write all of its fields to a string in
a dash-separated manner, and deserialize similarly.

The serializer and the implementation of the Identifiable interface are
the only changes needed to be able to store TimeControl data using our
repository system.