Adds Profile data class and Serializer
Profile stores information relevant to a specific player, namely their
name and preferred color. To conform to the Identifiable interface, the
appropriate properties are added.

To serialize a Profile, just write all of its fields to a string in
a dash-separated manner, and deserialize similarly. The color is stored
as its color int (a hex value) and read with the parseColor function.

The serializer and the implementation of the Identifiable interface are
the only changes needed to be able to store Profile data using our
repository system.