Adds Clock and ClockSet data class and Serializer
Clock stores information relevant to the timing information of a
specific player, namely their time control, the time they have left and
the start time of their last session (so they can reset back to it), and
of course the profile of the player.

ClockSet stores information relevant to a game, with one Clock for each
player, namely the name of the set / game, the Clocks of each player,
and what the current running clock is. To conform to the Identifiable
interface, the appropriate properties are added.

To serialize a Clock, just write all of its fields to a string in a
dash-separated manner, and deserialize similarly. The Profile and
TimeControl of the Clock are serialized just by their id. This means
that during deserialization only partial Profile and TimeControl data is
retrieved. To get the full Clock, data is needed from the Profile and
TimeControl repositories.

To serialize a ClockSet, just write all of its fields to a string in a
dash-separated manner, and deserialize similarly. The individual Clocks
of the ClockSet are also serialized and appended to the string.

The serializer and the implementation of the Identifiable interface are
the only changes needed to be able to store Clock and ClockSet data
using our repository system.