Improves EditClockSet screen
This screen now additionally lists the ClockSet's clocks (their profile
and time control) and allows the user to change the profile and time
control of each clock individually. To do that, the ListDataScreen
composable is used to list the available profile and time control data,
and to let the user pick one.

This change allows the EditClockSet screen to be used to also create
ClockSet data.

Cleans up ListDataScreen composable and removes the special case for the
CLOCK_SET DataType (the one which didn't allow it to be created).

Changelist:
- Updates EditClockSet composable
- Updates ListDataScreen composable
- Updates the parameters passed to EditClockSetScreena and
  ListDataScreen to see the changes