Creates separate ListClockSet composable
This composable is responsible for displaying ClockSet data, and allows
editing thereof. It is the entry point for the other List and EditXXX
composables, and therefore takes a ChckmViewModel as a parameter. The
EditProfile/TimeControl screens are accessed through the EditClockSet
composable, through which the user selects and edits individual Profile
and TimeControl entities. As the root of the hierarchy, it is
responsible for defining and passing down all the necessary callbacks.

Changelist:
- Creates ListClockSet
- Makes ListScreenContent composable public, to be used in ListClockSet
- Removes CLOCK_SET from ListType and ListScreenEnum, since it now has
  its own dedicated ListClockSet screen
- Adds LIST_CLOCKSET Screen constant to MainActivity, along with the
  three repositories, the viewModel, and the ListClockSetScreen