Adds MainScreen composable
MainScreen represents the main screen of the application, where the
current ClockSet is displayed, with the current player's clock at the
center. The players can navigate to the next and previous player, pause
and un-pause the current player's clock, reset and flag it, as well as
load a different game.

The MainActivity file is updated with more test data and the new
Screen.MAIN constant for the MainScreen composable.