ListScreen now shows the data sorted
Interface Displayable now extends Comparable, and each of the model data
classes implement the compareTo() method to sort data of the same type.

When sorting names, the case is not taken into account and when sorting
TimeControl data, sort first on time, and then on increment as a
tie-breaker.

Although the method signature allows it, comparing data of different
type (e.g. Profile to TimeControl) will throw an
IllegalArgumentException.