Fixes bug where editing time broke reset all
Editing the time of the current player now only changes the timeLeft and
their increment, leaving the underlying TimeControl data unaffected.

Previouisly, the changed TimeControl data would interfere with the reset
all functionality, meaning that reset all would reset to that edited
time, while the intended functionality for editing is that of a
temporary change.