Adds and uses ChckmScaffold composable
Adds a screen title and a back button, to enhance the user experience.
The title helps with navigation and keeping track of where the user is
at any point, and the back button (does nothing for now), allows the
user to go back to the previous screen without performing any action,
while previously there was no way to cancel and go back from the edit
profile screen.