Improves Color serialization
Now just writes an RRGGBB string, taking care to pad with '0's to ensure
a full-length RRGGBB string (the parseColor function throws otherwise).
This should (hopefully) cut down on weird edge cases (e.g. when
experimenting with Color.Unspecified, which has leading '0's and caused
the written string to be just "#0" instead of the full "#00000000").