Adds shared time control option for clock set
It is now possible for all players to share the same time control, by
choosing it once and having it apply to every player. This is controlled
by a switch. If all the players have the same time control when editing
a clock set, it is detected and the switch is set to true.